bugfix
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk/classes@3274 06c8f799-c779-4783-8af7-179afedb67c8

+++ java/lang/Math.java
@@ -249,7 +249,7 @@ public final class Math
 		if (x > 710)
 			return Double.POSITIVE_INFINITY;
 
-		double k = (int)(x / LN2);
+		int k = (int)(x / LN2);
 		if (x < 0)
 			k--;
 		x -= k * LN2;
